+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## v3.8.22
 
-- Introduce `fossa-deps-file` to specify custom fossa-deps file ([#1303](https://github.com/fossas/fossa-cli/pull/1303))
+- `fossa-deps`: `--fossa-deps-file` to specify custom fossa-deps file ([#1303](https://github.com/fossas/fossa-cli/pull/1303))
 
 ## v3.8.21